+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All documentation about Filebeat can be found here.
 ### Backward Compatibility Breaks
 
 ### Bugfixes
+- Omit 'fields' from event JSON when null.
 
 ### Added